Implement getPublicKeyIfApproved() call for better UX.
Background here: https://github.com/nostr-protocol/nips/pull/1599

The main idea is to simplify login logic in the client.
The client doesn't need to store the pubkey in localstorage,
it can query the extension for the currently used pubkey.

It also allows for the client to detect any changes in the extension
(using a different profile, revoking permission, etc)
With the localstorage solution it is ususally detected too late,
when the user wants to sign an event, but the pubkeys don't match.