Auto-set nostr.pubkey on page load for better UX.
Background here: https://github.com/nostr-protocol/nips/pull/1599

The main idea is to simplify login logic in the client.
The client doesn't need to store the pubkey in localstorage,
it can read it from nostr.pubkey, which is set by the extension
to the authorized pubkey.

It also allows for the client to detect any changes in the extension
(using a different profile, revoking permission, etc)
With the localstorage solution it is ususally detected too late,
when the user wants to sign an event, but the pubkeys don't match.